requirements.txt
python-docx
lxml
sentence-transformers
nltk
numpy
scipy
